# Let interfaces override __adapt__