… CHANGES.rst
… docs/api/declarations.rst
… src/zope/interface/tests/test_adapter.py
… src/zope/interface/tests/test_interface.py