Fixes #3 with no performance penalty (compared to always calling `self.__adapt__`) if this feature is not used. If it is, there's a minimum ~20% penalty, but that's still a lot faster than 4.7.2 (details in #3).

| Benchmark                                                   | Master | self.__adapt()                 | This PR             |
|-------------------------------------------------------------|----------------|------------------------------|------------------------------|
| call interface (provides; deep)                             | 369 ns         | 454 ns: 1.23x slower (+23%)  | not significant              |
| call interface (provides; wide)                             | 373 ns         | 457 ns: 1.22x slower (+22%)  | 365 ns: 1.02x faster (-2%)   |
| call interface (no alternate, no conform, not provided)     | 671 ns         | 760 ns: 1.13x slower (+13%)  | 636 ns: 1.06x faster (-5%)   |
| call interface (alternate, no conform, not provided)        | 395 ns         | 494 ns: 1.25x slower (+25%)  | not significant              |
| call interface (no alternate, valid conform, not provided)  | 250 ns         | not significant              | 227 ns: 1.10x faster (-9%)   |
| call interface (alternate, invalid conform, not provided)   | 348 ns         | 424 ns: 1.22x slower (+22%)  | not significant              |

Built on #201 because it touches the same C code; only the last commit is relevant for this PR.

Also some doc improvements in individual commits (`commits[-3:-1]` 😄 ) before I added the documentation for this feature.