@mgedmin 2020-04-06
Will this conflict with #201 much?  I'd rather postpone my review until after you rebase.

(I'm also a bit dubious on the feature itself.  Why would anyone want to override `Interface.__adapt__`?  PEP 246 was rejected in any case.)