@jamadden 2020-04-06
> Will this conflict with #201 much? I'd rather postpone my review until after you rebase.

No, because this is based on #201. Only the last commit here is relevant. I'll shortly hit the merge button over there and a lot of the commits here should just vanish.

> I'm also a bit dubious on the feature itself. Why would anyone want to override `Interface.__adapt__`?

🤷‍♂ The documentation basically claimed that you could do that, but it only worked correctly using the Python implementation. I don't have an immediate use-case myself, but I have found `__conform__` to be useful in the past and I can imagine that the…inversion?…of that in `__adapt__` could also be useful.